RealdataEXP/requirements.txt
boto3==1.28.62
botocore==1.31.62
bottle==0.12.21
contourpy==1.3.0
cycler==0.12.1
dbus-python==1.2.18
defusedxml==0.7.1
distro==1.5.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
future==0.18.3
Glances==3.3.1
gpg==1.15.1
idna==2.10
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==0.9.4
joblib==1.5.1
kiwisolver==1.4.7
libcomps==0.1.18
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==20.9
pandas==2.3.1
pillow==11.3.0
psutil==5.8.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==5.4.1
rpm==4.16.1.3
s3transfer==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
selinux==3.6
sepolicy==3.6
setools==4.4.4
six==1.15.0
SSSDConfig==2.9.5
sympy==1.14.0
systemd-python==234
threadpoolctl==3.6.0
torch==2.7.1
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.8.0
urllib3==1.26.5
zipp==3.23.0
